requirements.txt
textblob
flask
flask_cors
python-dotenv